Addition: How to play and Difficulty selector
I have added a slide down how to play box which details instructions on how to play. I have also added a radio button for selecting difficulty. I will add styling to the radio buttons in future and have it so when the quiz page opens, it displays your chosen difficulty